\name{mixor-deprecated}
\alias{mixor-deprecated}
\docType{package}
\title{
Deprecated Functions in Package mixor
}
\description{
These functions are provided for compatibility with older versions
  of \code{mixor} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{mixord: \code{\link{mixor}}}
  }
}

\author{
Kellie J. Archer, Donald Hedeker, Rachel Nordgren, Robert D. Gibbons, 
Maintainer: Kellie J. Archer <kjarcher@vcu.edu>
}

