\name{mixord}
\alias{mixord}
\title{
Old Function Name Replaced with mixor
}
\description{
The \code{mixord} function has been deprecated. Please use the \code{mixor} function for fitting longitudinal/clustered ordinal response models 
}
\author{
Kellie J. Archer, Donald Hedeker, Rachel Nordgren, Robert D. Gibbons
}
\seealso{
See Also as \code{\link{mixor}}, \code{\link{summary.mixor}}
}

